requirements.txt
flask
twilio
textblob
httpx
python-dotenv
streamlit
pandas
